+++ apps/dashboard/src/app/account/account-home/account-home.component.ts
@@ -1,11 +1,22 @@
-import { ChangeDetectionStrategy, ChangeDetectorRef, Component } from '@angular/core';
-import { Router } from '@angular/router';
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnInit } from '@angular/core';
+import { ActivatedRoute, Router } from '@angular/router';
 import { NavigationService, SDKService } from '@flaps/core';
 import { IKnowledgeBoxItem, StatsRange, StatsType } from '@nuclia/core';
-import { BehaviorSubject, catchError, combineLatest, map, Observable, of, shareReplay, switchMap, take } from 'rxjs';
+import {
+  BehaviorSubject,
+  catchError,
+  combineLatest,
+  filter,
+  map,
+  Observable,
+  of,
+  shareReplay,
+  switchMap,
+  take,
+} from 'rxjs';
 import { format } from 'date-fns';
 import { TranslateService } from '@ngx-translate/core';
-import { SisToastService } from '@nuclia/sistema';
+import { SisModalService, SisToastService } from '@nuclia/sistema';
 import { TickOptions } from '@flaps/common';
 import { ChartData, MetricsService } from '../metrics.service';
 
@@ -15,7 +26,7 @@ import { ChartData, MetricsService } from '../metrics.service';
   styleUrls: ['./account-home.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class AccountHomeComponent {
+export class AccountHomeComponent implements OnInit {
   account$ = this.metrics.account$;
   canUpgrade = this.metrics.canUpgrade;
   isSubscribed = this.metrics.isSubscribed;
@@ -82,7 +93,21 @@ export class AccountHomeComponent {
     private navigation: NavigationService,
     private router: Router,
     private metrics: MetricsService,
+    private route: ActivatedRoute,
+    private modal: SisModalService,
   ) {}
+
+  ngOnInit() {
+    this.route.queryParams
+      .pipe(
+        take(1),
+        filter((params) => params['setup'] === 'invite-collaborators'),
+      )
+      .subscribe((params) => {
+        // TODO: invite collaborators modal
+        // this.modal.openModal();
+      });
+  }
 
   toggleCharts() {
     this.allCharts = !this.allCharts;