+++ libs/sistema/src/index.ts
@@ -1,5 +1,6 @@
 export * from './lib/back-button';
 export * from './lib/dropdown-button';
+export * from './lib/help-card';
 export * from './lib/home-container';
 export * from './lib/icons';
 export * from './lib/json-viewer';